flask_s3_static_display_img_app/requirements.txt
Flask==1.0.2
boto3
Jinja2
datetime
python-dateutil
